Adds EditRoom and List-returns
Adds EditRoom method at all stages, updates previous room-getting methods to be one unified method with bool filtration and a listed return, and a get-room-by-id method for getting a singular room return by ID number alongside other minor fixes.

+++ src/models/Room.java
@@ -9,10 +9,10 @@ public class Room
     private BigDecimal price;
     private String room_type;
 
-    public Room(int id, boolean is_available, double price, String room_type) {
+    public Room(int id, boolean is_available, BigDecimal price, String room_type) {
         this.id = id;
         this.is_available = is_available;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
         this.room_type = room_type;
     }
